requirements.txt
# A.E.G.I.S. DEPENDENCY MANIFEST
# PROFILE: FORD 6.7L FORENSIC SUITE

# --- CORE HARDWARE INTERFACE ---
# 'obd' handles the ELM327/STN command set for the Ford 6.7L
obd==0.7.1
pyserial==3.5

# --- NUMERICS & PHYSICS ---
# Required for Zeldovich calculations and FFTs in temporal.py
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# --- ARTIFICIAL INTELLIGENCE ---
# Required for src/ai/meta_classifier.py and tools/train_dummy_model.py
tensorflow>=2.10.0
scikit-learn>=1.0.2

# --- SYSTEM & DASHBOARD ---
# 'colorama' ensures the Red/Green dashboard alerts work on Windows laptops
pyyaml>=6.0
colorama>=0.4.6
rich>=12.0.0

# --- DEVELOPMENT (OPTIONAL) ---
# For unit testing the physics engine
pytest>=7.0.0